pyproject.toml
[tool.poetry]
name = "mathsoc_bot"
version = "0.1.0"
description = "Mathsoc's discord bot"
authors = ["Ben Simms <ben@bensimms.moe>"]

[tool.poetry.dependencies]
python = "^3.8"
itsdangerous = "^1.1"
cryptography = "^3.0"
gino = "^1.0"
python-dotenv = "^0.14"
sqlalchemy_utils = "^0.36"
asyncpg = "^0.20"
alembic = "^1.4"
python-dateutil = "^2.8"
parsedatetime = "^2.6"
"discord.py" = "^1.4"
psycopg2-binary = "^2.8.5"

[tool.poetry.scripts]
start_bot = 'mathsoc_bot:run'
gen_tokens = 'mathsoc_bot:gen_tokens'

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.core.masonry.api"
